ipset: remove unneeded code
The family option only needs to be set during IPSET_CMD_CREATE, after
that it cannot be changed. See also lib/PROTOCOL in the ipset sources
(I consulted ipset 6.29).

The type parameter is not needed for the IPSET_CMD_ADD, IPSET_CMD_DEL
and IPSET_CMD_TEST commands. No need to check for it, if the setname
does not exist the command will fail anyway.